MB-64472: Fix TLS bad certificate error
When we toggle from "strict" Node-to-Node Encryption
to "control"/"all" setting which changes `DisableNonSSLPorts`
parameter, our HTTP server doesn't shutdown properly.
This causes `address already in use` error, which makes the
Eventing producer to restart unnecessarily. Additionally,
passing a node certificate in TLS config of http client
causes a "bad certificate" error because the server expects
a valid client certificate for authentication, not a
node certificate.

The fix in this patch include:

* Maintain references of `httpServer` and `httpSSLServer`
as function parameters during function recursion to
shutdown the servers correctly.

* Pass only `RootCAs` in the TLS config of clients by default,
which is required for verifying server certificates.

Change-Id: I4a73a1eded2780bed85148c806dec646ee0faa56
Reviewed-on: https://review.couchbase.org/c/eventing/+/220474
Tested-by: Eventing CI Bot
Reviewed-by: <ankit.prabhu@couchbase.com>
Reviewed-by: Rishit Chaudhary <rishit.chaudhary@couchbase.com>