MB-64952: Fix field names in the Eventing service's API responses
The field `function_scope` must have an object value with the fields
`bucket` and `scope` to ensure backwards compatibility with the old
Eventing architecture.
Additionally, this ensures that the Eventing service's responses sent
for requests made to the GET `/api/v1/status/{function}` REST API
endpoint is correct as per the ClusterRebalanceIT analytics test.

Change-Id: Ia941b76e9d312e0c2e91f561a6c4ffad232387a7
Reviewed-on: https://review.couchbase.org/c/eventing/+/222658
Reviewed-by: <ankit.prabhu@couchbase.com>
Tested-by: Rishit Chaudhary <rishit.chaudhary@couchbase.com>