MB-67055: Handle `bucket.scope.*` keyspace permissions correctly in `getUserInfo` endpoint
Bug: We iterate on bucket, scope, and collections in nested
for-loops while changing value of keyspace. During the next
iteration of the scope for-loop, the value of collection from
the prev iteration is retained, leading to incorrect keyspace output.

Key fix: Reset collection name to "*" (wildcard) at the start of
scope for-loop to avoid retaining collection value from prev iteration.

Additional changes:
- Added utility function `SortKeyspaces` to sort keyspaces
based on BucketName, ScopeName, and CollectionName, ensuring
that the GlobalValue is prioritized. This function will
help return the keyspaces in a sorted format consistently.

- Added helper function `checkAndAppendPermissions` for cleaner
code and improved readability.

Change-Id: I5f917e9f5cfa1342a5ada76451104934cd242982
Reviewed-on: https://review.couchbase.org/c/eventing/+/228841
Reviewed-by: Rishit Chaudhary <rishit.chaudhary@couchbase.com>
Reviewed-by: <ankit.prabhu@couchbase.com>
Tested-by: Hiren <hiren.bavaskar@couchbase.com>